+++ project/src/components/common/price.tsx
@@ -5,9 +5,9 @@ type PriceProps = {
 
 function Price({ price, uniqUrl}: PriceProps): JSX.Element {
   return (
-    <div className={ uniqUrl ? 'property__price' : 'place-card__price'} >
-      <b className={ uniqUrl ? 'property__price-value' : 'place-card__price-value'} >&euro;{ price }</b>
-      <span className={ uniqUrl ? 'property__price-text' : 'place-card__price-text'} >&nbsp;night</span>
+    <div className={ `${uniqUrl ? 'property' : 'place-card'}__price` }>
+      <b className={ `${uniqUrl ? 'property' : 'place-card'}__price-value` }>&euro;{ price }</b>
+      <span className={ `${uniqUrl ? 'property' : 'place-card'}__price-text` }>&nbsp;night</span>
     </div>
   );
 }
